move indexes to own directory
