+++ conftest.py
@@ -8,7 +8,7 @@ from testing import (
     WIGGLE_NUMBER_TEST_FILE_PATH,
 )
 from wiggle_query_language.graph.database.database import wipe_database
-from wiggle_query_language.graph.database.relationship_index import (
+from wiggle_query_language.graph.database.indexes.relationship_index import (
     wipe_relationship_index,
 )
