+++ wiggle_query_language/clauses/make/make.py
@@ -15,7 +15,7 @@ from wiggle_query_language.clauses.parsing_helpers.parse_properties import (
     get_property_dict,
 )
 from wiggle_query_language.graph.database.database import add_item_to_database
-from wiggle_query_language.graph.database.relationship_index import (
+from wiggle_query_language.graph.database.indexes.relationship_index import (
     add_items_to_relationship_index,
 )
 from wiggle_query_language.graph.state.wiggle_number import (
